requirements.txt
flask
APScheduler
requests
python-dotenv
google-generativeai
